Moved getWorldPosition method to position component file. Maybe will do the same for client.
